+++ Frends.CSV.Create/Frends.CSV.Create/Definitions/Result.cs
@@ -9,17 +9,16 @@ public class Result
     /// Operation complete without errors.
     /// </summary>
     /// <example>true</example>
-    public bool Success { get; private set; }
+    public bool Success { get; init; }
 
     /// <summary>
     /// CSV string.
     /// </summary>
     /// <example>First;Second;"foo";"bar"</example>
-    public string CSV { get; private set; }
+    public string CSV { get; init; }
 
-    internal Result(bool success, string cSV)
-    {
-        Success = success;
-        CSV = cSV;
-    }
+    /// <summary>
+    /// Error information. Null when Success is true.
+    /// </summary>
+    public Error Error { get; init; }
 }